refactor: getRandomNumbers 함수 유틸리티 함수로 변경
